[HUDI-7111] Fix performance regression of tag when written into simple bucket index table (#10130)
