+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/index/bucket/HoodieBucketIndex.java
@@ -19,13 +19,9 @@
 package org.apache.hudi.index.bucket;
 
 import org.apache.hudi.client.WriteStatus;
-import org.apache.hudi.client.utils.LazyIterableIterator;
 import org.apache.hudi.common.data.HoodieData;
 import org.apache.hudi.common.engine.HoodieEngineContext;
-import org.apache.hudi.common.model.HoodieRecord;
-import org.apache.hudi.common.model.HoodieRecordLocation;
 import org.apache.hudi.common.model.WriteOperationType;
-import org.apache.hudi.common.util.Option;
 import org.apache.hudi.config.HoodieWriteConfig;
 import org.apache.hudi.exception.HoodieIndexException;
 import org.apache.hudi.index.HoodieIndex;
@@ -36,8 +32,6 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Arrays;
 import java.util.List;
-
-import static org.apache.hudi.index.HoodieIndexUtils.tagAsNewRecordIfNeeded;
 
 /**
  * Hash indexing mechanism.
@@ -63,30 +57,6 @@ public abstract class HoodieBucketIndex extends HoodieIndex<Object, Object> {
                                                 HoodieTable hoodieTable)
       throws HoodieIndexException {
     return writeStatuses;
-  }
-
-  @Override
-  public <R> HoodieData<HoodieRecord<R>> tagLocation(
-      HoodieData<HoodieRecord<R>> records, HoodieEngineContext context,
-      HoodieTable hoodieTable)
-      throws HoodieIndexException {
-    // Get bucket location mapper for the given partitions
-    List<String> partitions = records.map(HoodieRecord::getPartitionPath).distinct().collectAsList();
-    LOG.info("Get BucketIndexLocationMapper for partitions: " + partitions);
-    BucketIndexLocationMapper mapper = getLocationMapper(hoodieTable, partitions);
-
-    return records.mapPartitions(iterator ->
-        new LazyIterableIterator<HoodieRecord<R>, HoodieRecord<R>>(iterator) {
-          @Override
-          protected HoodieRecord<R> computeNext() {
-            // TODO maybe batch the operation to improve performance
-            HoodieRecord record = inputItr.next();
-            Option<HoodieRecordLocation> loc = mapper.getRecordLocation(record.getKey());
-            return tagAsNewRecordIfNeeded(record, loc);
-          }
-        },
-        false
-    );
   }
 
   @Override
@@ -127,9 +97,4 @@ public abstract class HoodieBucketIndex extends HoodieIndex<Object, Object> {
   public int getNumBuckets() {
     return numBuckets;
   }
-
-  /**
-   * Get a location mapper for the given table & partitionPath
-   */
-  protected abstract BucketIndexLocationMapper getLocationMapper(HoodieTable table, List<String> partitionPath);
 }
